refactor(frontend): 巨大単一ファイル分割 + routing 抽出 (#1145 Phase-5+6+7)
#1145 シリーズ最終 Phase の 3 件を 1 PR にまとめて実施 (鉄則 3 同根統合)。

Phase-5 (ConventionsCatalogView 分割):
- 72KB / 1982 行 → 28KB / 619 行 (-61%、-44KB)
- 13 category 編集 panel (`internal/categories/MsgPanel.tsx` ほか) + 1 read-only ExtensionCategoriesPanel = 14 panel に分離
- 共通部品を `internal/SharedRowParts.tsx` / `internal/sharedTypes.ts` / `internal/sharedOptions.ts` に集約
- 新規 rendering test 20 件 (conventions/ test 0 件 → 20 件)

Phase-6 (ScreenItemsView 分割):
- 64KB / 1545 行 → 56KB / 1278 行 (-12%、-8KB)
- 出力設定 sub-form OutputFields を `internal/sections/OutputFields.tsx` に抽出
- 共通定数を `internal/screenItemsConstants.ts` に分離
- 新規 test 15 件 (screen-items/ test 0 件 → 15 件)
- 残 KB は state-coupled な ID 編集 / lint / select handler 群で、機械的抽出は props inflation を招くため対象外と判断

Phase-7 (AppShell routing 抽出):
- AppShell.tsx 984 → 955 行 (-29 行) + 純粋関数 `evaluateRoutingGuard` を `routing/workspaceRouting.ts` へ抽出
- 旧 test ファイル内に inline 重複定義していた `wsPath` / `evaluateRoutingGuard` を import 参照に切替 (逆コロケーション解消)
- `useWorkspacePath` hook も内部で `wsPath` 純粋関数を参照する形に整理
- test 19 件 (旧 14 件 + 新 5 件: 空文字 wsId / wsId が recent にある復元 + 各 navigate 経路)

検証:
- `npm run build` pass
- `npm run test:unit -- --run` pass (169 file / 2302 test、+38 件追加)
- `npm run lint` pass (5 件 pre-existing warning のみ)

Closes #1145
