+++ frontend/src/components/conventions/ConventionsCatalogView.tsx
@@ -6,8 +6,12 @@
  *   - 入力バリデーション: msg / regex / limit
  *   - 役割・権限: role / permission
  *   - プロダクト規約: scope / currency / tax / auth / db / numbering / tx / externalOutcomeDefaults
+ *
+ * #1145 Phase-5: 各カテゴリの編集 panel を `internal/categories/*Panel.tsx` に抽出。
+ * 共通部品は `internal/SharedRowParts.tsx` / `internal/sharedOptions.ts` / `internal/sharedTypes.ts` に集約。
+ * 拡張カテゴリ read-only 表示は `internal/categories/ExtensionCategoriesPanel.tsx` に分離。
  */
-import { Fragment, useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { TableSubToolbar } from "../table/TableSubToolbar";
 import { EditorHeader } from "../common/EditorHeader";
 import { ServerChangeBanner } from "../common/ServerChangeBanner";
@@ -34,22 +38,25 @@ import {
   type ConventionIssue,
   checkConventionsCatalogIntegrity,
 } from "../../schemas/conventionsValidator";
-import type {
-  ScopeEntry,
-  CurrencyEntry,
-  TaxEntry,
-  AuthEntry,
-  RoleEntry,
-  PermissionEntry,
-  DbEntry,
-  NumberingEntry,
-  TxEntry,
-  ExternalOutcomeEntry,
-  SemVer,
-} from "../../types/v3";
+import type { SemVer } from "../../types/v3";
 import { mcpBridge } from "../../mcp/mcpBridge";
 import "../../styles/conventions.css";
 import "../../styles/editMode.css";
+
+import { MsgPanel } from "./internal/categories/MsgPanel";
+import { RegexPanel } from "./internal/categories/RegexPanel";
+import { LimitPanel } from "./internal/categories/LimitPanel";
+import { ScopePanel } from "./internal/categories/ScopePanel";
+import { CurrencyPanel } from "./internal/categories/CurrencyPanel";
+import { TaxPanel } from "./internal/categories/TaxPanel";
+import { AuthPanel } from "./internal/categories/AuthPanel";
+import { RolePanel } from "./internal/categories/RolePanel";
+import { PermissionPanel } from "./internal/categories/PermissionPanel";
+import { DbPanel } from "./internal/categories/DbPanel";
+import { NumberingPanel } from "./internal/categories/NumberingPanel";
+import { TxPanel } from "./internal/categories/TxPanel";
+import { ExternalOutcomeDefaultsPanel } from "./internal/categories/ExternalOutcomeDefaultsPanel";
+import { ExtensionCategoriesPanel } from "./internal/categories/ExtensionCategoriesPanel";
 
 type Category =
   | "msg" | "regex" | "limit"
@@ -471,7 +478,7 @@ export function ConventionsCatalogView() {
 
       <div className="conventions-category-content">
         {activeCategory === "msg" && (
-          <MsgEditor
+          <MsgPanel
             msg={catalog.msg ?? {}}
             onAdd={addMsg}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.msg?.[key]) Object.assign(c.msg[key], patch); })}
@@ -481,7 +488,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "regex" && (
-          <RegexEditor
+          <RegexPanel
             regex={catalog.regex ?? {}}
             onAdd={addRegex}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.regex?.[key]) Object.assign(c.regex[key], patch); })}
@@ -491,7 +498,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "limit" && (
-          <LimitEditor
+          <LimitPanel
             limit={catalog.limit ?? {}}
             onAdd={addLimit}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.limit?.[key]) Object.assign(c.limit[key], patch); })}
@@ -501,7 +508,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "scope" && (
-          <ScopeEditor
+          <ScopePanel
             scope={catalog.scope ?? {}}
             onAdd={addScope}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.scope?.[key]) Object.assign(c.scope[key], patch); })}
@@ -511,7 +518,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "currency" && (
-          <CurrencyEditor
+          <CurrencyPanel
             currency={catalog.currency ?? {}}
             onAdd={addCurrency}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.currency?.[key]) Object.assign(c.currency[key], patch); })}
@@ -521,7 +528,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "tax" && (
-          <TaxEditor
+          <TaxPanel
             tax={catalog.tax ?? {}}
             onAdd={addTax}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.tax?.[key]) Object.assign(c.tax[key], patch); })}
@@ -531,7 +538,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "auth" && (
-          <AuthEditor
+          <AuthPanel
             auth={catalog.auth ?? {}}
             onAdd={addAuth}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.auth?.[key]) Object.assign(c.auth[key], patch); })}
@@ -541,7 +548,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "role" && (
-          <RoleEditor
+          <RolePanel
             role={catalog.role ?? {}}
             permissionKeys={Object.keys(catalog.permission ?? {})}
             issues={integrityIssues}
@@ -553,7 +560,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "permission" && (
-          <PermissionEditor
+          <PermissionPanel
             permission={catalog.permission ?? {}}
             onAdd={addPermission}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.permission?.[key]) Object.assign(c.permission[key], patch); })}
@@ -563,7 +570,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "db" && (
-          <DbEditor
+          <DbPanel
             db={catalog.db ?? {}}
             onAdd={addDb}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.db?.[key]) Object.assign(c.db[key], patch); })}
@@ -573,7 +580,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "numbering" && (
-          <NumberingEditor
+          <NumberingPanel
             numbering={catalog.numbering ?? {}}
             onAdd={addNumbering}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.numbering?.[key]) Object.assign(c.numbering[key], patch); })}
@@ -583,7 +590,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "tx" && (
-          <TxEditor
+          <TxPanel
             tx={catalog.tx ?? {}}
             onAdd={addTx}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => { if (c.tx?.[key]) Object.assign(c.tx[key], patch); })}
@@ -593,7 +600,7 @@ export function ConventionsCatalogView() {
           />
         )}
         {activeCategory === "externalOutcomeDefaults" && (
-          <ExternalOutcomeDefaultEditor
+          <ExternalOutcomeDefaultsPanel
             entries={catalog.externalOutcomeDefaults ?? {}}
             onAdd={addExternalOutcomeDefault}
             onUpdate={(key, patch) => updateSilentWithDraft((c) => {
@@ -608,1375 +615,5 @@ export function ConventionsCatalogView() {
 
       <ExtensionCategoriesPanel extensionCategories={catalog.extensionCategories} />
     </div>
-  );
-}
-
-/** 拡張カテゴリ表示 (read-only)。extensions.v3 の conventionCategories で定義された業界規約を一覧表示。 */
-function ExtensionCategoriesPanel({
-  extensionCategories,
-}: {
-  extensionCategories?: Record<string, Record<string, unknown>>;
-}) {
-  const entries = Object.entries(extensionCategories ?? {});
-  return (
-    <section className="conventions-extension-categories">
-      <h3 className="conventions-section-title">
-        <i className="bi bi-puzzle" /> 拡張カテゴリ
-        <small className="text-muted ms-2">
-          (extensions.v3 の conventionCategories で定義、`@conv.&lt;categoryName&gt;.&lt;key&gt;` で参照)
-        </small>
-      </h3>
-      {entries.length === 0 ? (
-        <div className="conventions-empty">拡張カテゴリは定義されていません。</div>
-      ) : (
-        <table className="conventions-table">
-          <thead>
-            <tr>
-              <th>カテゴリ名</th>
-              <th>エントリ数</th>
-              <th>キー一覧</th>
-            </tr>
-          </thead>
-          <tbody>
-            {entries.map(([catName, catEntries]) => {
-              const keys = Object.keys(catEntries ?? {});
-              return (
-                <tr key={catName}>
-                  <td><code className="conventions-key-badge">@conv.{catName}.*</code></td>
-                  <td>{keys.length}</td>
-                  <td className="text-muted">
-                    {keys.length > 0 ? keys.join(", ") : "—"}
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      )}
-    </section>
-  );
-}
-
-// ── 共通部品 ────────────────────────────────────────────────────────────
-
-function DeleteBtn({ onClick, isReadonly }: { onClick: () => void; isReadonly?: boolean }) {
-  return (
-    <button
-      type="button"
-      className="btn btn-sm btn-link text-danger p-0"
-      onClick={onClick}
-      title="削除"
-      aria-label="削除"
-      disabled={isReadonly}
-    >
-      <i className="bi bi-x" />
-    </button>
-  );
-}
-
-function NewKeyRow({
-  placeholder, value, setValue, onAdd, disabled, isReadonly,
-}: {
-  placeholder: string;
-  value: string;
-  setValue: (v: string) => void;
-  onAdd: () => void;
-  disabled: boolean;
-  isReadonly?: boolean;
-}) {
-  return (
-    <div className="conventions-new-key-row">
-      <input
-        className="form-control form-control-sm conventions-new-key-input"
-        placeholder={placeholder}
-        value={value}
-        onChange={(e) => setValue(e.target.value)}
-        onKeyDown={(e) => { if (e.key === "Enter" && !disabled && !isReadonly) { e.preventDefault(); onAdd(); } }}
-        disabled={isReadonly}
-      />
-      <button
-        type="button"
-        className="btn btn-sm btn-outline-primary"
-        onClick={onAdd}
-        disabled={disabled || isReadonly}
-      >
-        <i className="bi bi-plus-lg" /> 追加
-      </button>
-    </div>
-  );
-}
-
-function EntriesWrapper({ children, empty }: { children: React.ReactNode; empty: boolean }) {
-  return (
-    <div className="conventions-entries">
-      {empty && <div className="conventions-empty">エントリがありません。下の入力欄から追加してください。</div>}
-      {children}
-    </div>
-  );
-}
-
-// ── 既存 3 エディタ ─────────────────────────────────────────────────────
-
-interface MsgEntryLocal {
-  template: string;
-  params?: string[];
-  description?: string;
-}
-
-function MsgEditor({
-  msg, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  msg: Record<string, MsgEntryLocal>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<MsgEntryLocal>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(msg);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "14em" }} />
-          <col />
-          <col style={{ width: "18em" }} />
-          <col style={{ width: "18em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.msg.xxx)</th>
-            <th>template</th>
-            <th>params (カンマ区切り)</th>
-            <th>description</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.template}
-                  onChange={(e) => onUpdate(key, { template: e.target.value })}
-                  onBlur={onCommit}
-                  placeholder="{label}は必須入力です"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={(entry.params ?? []).join(", ")}
-                  onChange={(e) => {
-                    const params = e.target.value.split(",").map((s) => s.trim()).filter(Boolean);
-                    onUpdate(key, { params: params.length > 0 ? params : undefined });
-                  }}
-                  onBlur={onCommit}
-                  placeholder="label, max"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: required)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !msg[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(msg, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-interface RegexEntryLocal {
-  pattern: string;
-  flags?: string;
-  description?: string;
-}
-
-function RegexEditor({
-  regex, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  regex: Record<string, RegexEntryLocal>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<RegexEntryLocal>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(regex);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "14em" }} />
-          <col />
-          <col style={{ width: "6em" }} />
-          <col style={{ width: "18em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.regex.xxx)</th>
-            <th>pattern</th>
-            <th>flags</th>
-            <th>description</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  className="form-control form-control-sm conventions-mono"
-                  value={entry.pattern}
-                  onChange={(e) => onUpdate(key, { pattern: e.target.value })}
-                  onBlur={onCommit}
-                  placeholder="^[A-Za-z0-9]+$"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm conventions-mono"
-                  value={entry.flags ?? ""}
-                  onChange={(e) => onUpdate(key, { flags: e.target.value || undefined })}
-                  onBlur={onCommit}
-                  placeholder="i"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: phone-jp)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !regex[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(regex, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-interface LimitEntryLocal {
-  value: number;
-  unit?: string;
-  description?: string;
-}
-
-function LimitEditor({
-  limit, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  limit: Record<string, LimitEntryLocal>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<LimitEntryLocal>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(limit);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "14em" }} />
-          <col style={{ width: "10em" }} />
-          <col style={{ width: "8em" }} />
-          <col />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.limit.xxx)</th>
-            <th>value</th>
-            <th>unit</th>
-            <th>description</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  type="number"
-                  className="form-control form-control-sm"
-                  value={entry.value}
-                  onChange={(e) => onUpdate(key, { value: Number(e.target.value) })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.unit ?? ""}
-                  onChange={(e) => onUpdate(key, { unit: e.target.value || undefined })}
-                  onBlur={onCommit}
-                  placeholder="char"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: emailMax)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !limit[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(limit, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-// ── プロダクト規約 + role/permission エディタ群 ─────────────────────────
-
-function DefaultCell<T extends { default?: boolean }>({
-  entry, onUpdate, onCommit,
-}: {
-  entry: T;
-  onUpdate: (patch: Partial<T>) => void;
-  onCommit: () => void;
-}) {
-  return (
-    <td className="text-center" title="プロジェクト全体の ambient default として扱う">
-      <input
-        type="checkbox"
-        checked={entry.default ?? false}
-        onChange={(e) => {
-          onUpdate({ default: e.target.checked || undefined } as Partial<T>);
-          onCommit();
-        }}
-      />
-    </td>
-  );
-}
-
-function ScopeEditor({
-  scope, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  scope: Record<string, ScopeEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<ScopeEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(scope);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "14em" }} />
-          <col style={{ width: "16em" }} />
-          <col />
-          <col style={{ width: "5em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.scope.xxx)</th>
-            <th>value</th>
-            <th>description</th>
-            <th title="プロジェクト全体の ambient default として扱う">default</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.value}
-                  onChange={(e) => onUpdate(key, { value: e.target.value })}
-                  onBlur={onCommit}
-                  placeholder="domestic"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <DefaultCell
-                entry={entry}
-                onUpdate={(patch) => onUpdate(key, patch)}
-                onCommit={onCommit}
-              />
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: customerRegion)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !scope[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(scope, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-const ROUNDING_OPTIONS = ["", "floor", "ceil", "round"] as const;
-
-function CurrencyEditor({
-  currency, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  currency: Record<string, CurrencyEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<CurrencyEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(currency);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "12em" }} />
-          <col style={{ width: "7em" }} />
-          <col style={{ width: "8em" }} />
-          <col style={{ width: "10em" }} />
-          <col />
-          <col style={{ width: "5em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.currency.xxx)</th>
-            <th>code (ISO 4217)</th>
-            <th>subunit</th>
-            <th>roundingMode</th>
-            <th>description</th>
-            <th title="プロジェクト全体の ambient default として扱う">default</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  className="form-control form-control-sm conventions-mono"
-                  value={entry.code}
-                  onChange={(e) => onUpdate(key, { code: e.target.value })}
-                  onBlur={onCommit}
-                  placeholder="JPY"
-                />
-              </td>
-              <td>
-                <input
-                  type="number"
-                  className="form-control form-control-sm"
-                  value={entry.subunit ?? ""}
-                  onChange={(e) => onUpdate(key, { subunit: e.target.value === "" ? undefined : Number(e.target.value) })}
-                  onBlur={onCommit}
-                  placeholder="0"
-                  min={0}
-                />
-              </td>
-              <td>
-                <select
-                  className="form-select form-select-sm"
-                  value={entry.roundingMode ?? ""}
-                  onChange={(e) => {
-                    onUpdate(key, { roundingMode: (e.target.value || undefined) as CurrencyEntry["roundingMode"] });
-                    onCommit();
-                  }}
-                >
-                  {ROUNDING_OPTIONS.map((o) => <option key={o} value={o}>{o || "(未指定)"}</option>)}
-                </select>
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <DefaultCell
-                entry={entry}
-                onUpdate={(patch) => onUpdate(key, patch)}
-                onCommit={onCommit}
-              />
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: jpy)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !currency[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(currency, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-function TaxEditor({
-  tax, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  tax: Record<string, TaxEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<TaxEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(tax);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "12em" }} />
-          <col style={{ width: "10em" }} />
-          <col style={{ width: "10em" }} />
-          <col style={{ width: "10em" }} />
-          <col />
-          <col style={{ width: "5em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.tax.xxx)</th>
-            <th>kind</th>
-            <th>rate (0〜1)</th>
-            <th>roundingMode</th>
-            <th>description</th>
-            <th title="プロジェクト全体の ambient default として扱う">default</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <select
-                  className="form-select form-select-sm"
-                  value={entry.kind}
-                  onChange={(e) => { onUpdate(key, { kind: e.target.value as TaxEntry["kind"] }); onCommit(); }}
-                >
-                  <option value="exclusive">exclusive (外税)</option>
-                  <option value="inclusive">inclusive (内税)</option>
-                </select>
-              </td>
-              <td>
-                <input
-                  type="number"
-                  className="form-control form-control-sm"
-                  value={entry.rate}
-                  onChange={(e) => onUpdate(key, { rate: Number(e.target.value) })}
-                  onBlur={onCommit}
-                  step={0.01}
-                  min={0}
-                  max={1}
-                  placeholder="0.10"
-                />
-              </td>
-              <td>
-                <select
-                  className="form-select form-select-sm"
-                  value={entry.roundingMode ?? ""}
-                  onChange={(e) => {
-                    onUpdate(key, { roundingMode: (e.target.value || undefined) as TaxEntry["roundingMode"] });
-                    onCommit();
-                  }}
-                >
-                  {ROUNDING_OPTIONS.map((o) => <option key={o} value={o}>{o || "(未指定)"}</option>)}
-                </select>
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <DefaultCell
-                entry={entry}
-                onUpdate={(patch) => onUpdate(key, patch)}
-                onCommit={onCommit}
-              />
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: standard)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !tax[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(tax, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-function AuthEditor({
-  auth, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  auth: Record<string, AuthEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<AuthEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(auth);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "12em" }} />
-          <col style={{ width: "14em" }} />
-          <col style={{ width: "14em" }} />
-          <col style={{ width: "14em" }} />
-          <col />
-          <col style={{ width: "5em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.auth.xxx)</th>
-            <th>scheme</th>
-            <th>sessionStorage</th>
-            <th>passwordHash</th>
-            <th>description</th>
-            <th title="プロジェクト全体の ambient default として扱う">default</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.scheme}
-                  onChange={(e) => onUpdate(key, { scheme: e.target.value })}
-                  onBlur={onCommit}
-                  placeholder="session-cookie"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.sessionStorage ?? ""}
-                  onChange={(e) => onUpdate(key, { sessionStorage: e.target.value || undefined })}
-                  onBlur={onCommit}
-                  placeholder="httpOnly-cookie"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.passwordHash ?? ""}
-                  onChange={(e) => onUpdate(key, { passwordHash: e.target.value || undefined })}
-                  onBlur={onCommit}
-                  placeholder="bcrypt(cost=12)"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <DefaultCell
-                entry={entry}
-                onUpdate={(patch) => onUpdate(key, patch)}
-                onCommit={onCommit}
-              />
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: default)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !auth[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(auth, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-function RoleEditor({
-  role, permissionKeys, issues, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  role: Record<string, RoleEntry>;
-  permissionKeys: string[];
-  issues: ConventionIssue[];
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<RoleEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(role);
-  const roleKeys = Object.keys(role);
-
-  // role.<key> から始まる issue を key 別に集約
-  const issuesByKey = useMemo(() => {
-    const map = new Map<string, ConventionIssue[]>();
-    for (const iss of issues) {
-      const m = /^role\.([^.[]+)/.exec(iss.path);
-      if (!m) continue;
-      const k = m[1];
-      const arr = map.get(k) ?? [];
-      arr.push(iss);
-      map.set(k, arr);
-    }
-    return map;
-  }, [issues]);
-
-  const permissionListId = "conventions-permission-keys";
-  const roleListId = "conventions-role-keys";
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <datalist id={permissionListId}>
-        {permissionKeys.map((k) => <option key={k} value={k} />)}
-      </datalist>
-      <datalist id={roleListId}>
-        {roleKeys.map((k) => <option key={k} value={k} />)}
-      </datalist>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "12em" }} />
-          <col style={{ width: "10em" }} />
-          <col style={{ width: "14em" }} />
-          <col />
-          <col style={{ width: "16em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.role.xxx)</th>
-            <th>name</th>
-            <th>description</th>
-            <th>permissions (カンマ区切り)</th>
-            <th>inherits (カンマ区切り)</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => {
-            const rowIssues = issuesByKey.get(key) ?? [];
-            const hasPermIssue = rowIssues.some((i) =>
-              i.code === "UNKNOWN_CONV_ROLE_PERMISSION" && i.path.startsWith(`role.${key}.permissions`),
-            );
-            const hasInheritsIssue = rowIssues.some((i) =>
-              (i.code === "UNKNOWN_CONV_ROLE_INHERITS" || i.code === "ROLE_INHERITS_CYCLE") &&
-              i.path.startsWith(`role.${key}.inherits`),
-            );
-            return (
-              <Fragment key={key}>
-                <tr>
-                  <td><code className="conventions-key-badge">{key}</code></td>
-                  <td>
-                    <input
-                      className="form-control form-control-sm"
-                      value={entry.name ?? ""}
-                      onChange={(e) => onUpdate(key, { name: e.target.value || undefined })}
-                      onBlur={onCommit}
-                      placeholder="顧客"
-                    />
-                  </td>
-                  <td>
-                    <input
-                      className="form-control form-control-sm"
-                      value={entry.description ?? ""}
-                      onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                      onBlur={onCommit}
-                    />
-                  </td>
-                  <td>
-                    <input
-                      className={`form-control form-control-sm ${hasPermIssue ? "is-invalid" : ""}`}
-                      list={permissionListId}
-                      value={(entry.permissions ?? []).join(", ")}
-                      onChange={(e) => {
-                        const perms = e.target.value.split(",").map((s) => s.trim()).filter(Boolean);
-                        onUpdate(key, { permissions: perms });
-                      }}
-                      onBlur={onCommit}
-                      placeholder="order.create, order.read"
-                    />
-                  </td>
-                  <td>
-                    <input
-                      className={`form-control form-control-sm ${hasInheritsIssue ? "is-invalid" : ""}`}
-                      list={roleListId}
-                      value={(entry.inherits ?? []).join(", ")}
-                      onChange={(e) => {
-                        const inh = e.target.value.split(",").map((s) => s.trim()).filter(Boolean);
-                        onUpdate(key, { inherits: inh.length > 0 ? inh : undefined });
-                      }}
-                      onBlur={onCommit}
-                      placeholder="customer"
-                    />
-                  </td>
-                  <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-                </tr>
-                {rowIssues.length > 0 && (
-                  <tr className="conventions-row-issues">
-                    <td />
-                    <td colSpan={5}>
-                      <ul className="conventions-issue-list">
-                        {rowIssues.map((iss, i) => (
-                          <li key={i} className="conventions-issue">
-                            <i className="bi bi-exclamation-triangle-fill" />
-                            <span className="conventions-issue-message">{iss.message}</span>
-                          </li>
-                        ))}
-                      </ul>
-                    </td>
-                  </tr>
-                )}
-              </Fragment>
-            );
-          })}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: customer)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !role[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(role, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-const SCOPE_OPTIONS = ["", "all", "own", "department"] as const;
-
-function PermissionEditor({
-  permission, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  permission: Record<string, PermissionEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<PermissionEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(permission);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "14em" }} />
-          <col style={{ width: "12em" }} />
-          <col style={{ width: "10em" }} />
-          <col style={{ width: "10em" }} />
-          <col />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.permission.xxx)</th>
-            <th>resource</th>
-            <th>action</th>
-            <th>scope</th>
-            <th>description</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.resource}
-                  onChange={(e) => onUpdate(key, { resource: e.target.value })}
-                  onBlur={onCommit}
-                  placeholder="Order"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.action}
-                  onChange={(e) => onUpdate(key, { action: e.target.value })}
-                  onBlur={onCommit}
-                  placeholder="create"
-                />
-              </td>
-              <td>
-                <select
-                  className="form-select form-select-sm"
-                  value={entry.scope ?? ""}
-                  onChange={(e) => {
-                    onUpdate(key, { scope: (e.target.value || undefined) as PermissionEntry["scope"] });
-                    onCommit();
-                  }}
-                >
-                  {SCOPE_OPTIONS.map((o) => <option key={o} value={o}>{o || "(未指定)"}</option>)}
-                </select>
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: order.create)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !permission[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(permission, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-function DbEditor({
-  db, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  db: Record<string, DbEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<DbEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(db);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "12em" }} />
-          <col style={{ width: "12em" }} />
-          <col style={{ width: "10em" }} />
-          <col style={{ width: "16em" }} />
-          <col style={{ width: "12em" }} />
-          <col />
-          <col style={{ width: "5em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.db.xxx)</th>
-            <th>engine</th>
-            <th>namingConvention</th>
-            <th>timestampColumns (カンマ区切り)</th>
-            <th>logicalDeleteColumn</th>
-            <th>description</th>
-            <th title="プロジェクト全体の ambient default として扱う">default</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.engine ?? ""}
-                  onChange={(e) => onUpdate(key, { engine: e.target.value || undefined })}
-                  onBlur={onCommit}
-                  placeholder="postgresql@14"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.namingConvention ?? ""}
-                  onChange={(e) => onUpdate(key, { namingConvention: e.target.value || undefined })}
-                  onBlur={onCommit}
-                  placeholder="snake_case"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={(entry.timestampColumns ?? []).join(", ")}
-                  onChange={(e) => {
-                    const cols = e.target.value.split(",").map((s) => s.trim()).filter(Boolean);
-                    onUpdate(key, { timestampColumns: cols.length > 0 ? cols : undefined });
-                  }}
-                  onBlur={onCommit}
-                  placeholder="created_at, updated_at"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.logicalDeleteColumn ?? ""}
-                  onChange={(e) => onUpdate(key, { logicalDeleteColumn: e.target.value || undefined })}
-                  onBlur={onCommit}
-                  placeholder="is_deleted"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <DefaultCell
-                entry={entry}
-                onUpdate={(patch) => onUpdate(key, patch)}
-                onCommit={onCommit}
-              />
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: default)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !db[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(db, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-function NumberingEditor({
-  numbering, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  numbering: Record<string, NumberingEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<NumberingEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(numbering);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "14em" }} />
-          <col style={{ width: "14em" }} />
-          <col style={{ width: "18em" }} />
-          <col />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.numbering.xxx)</th>
-            <th>format</th>
-            <th>implementation</th>
-            <th>description</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <input
-                  className="form-control form-control-sm conventions-mono"
-                  value={entry.format}
-                  onChange={(e) => onUpdate(key, { format: e.target.value })}
-                  onBlur={onCommit}
-                  placeholder="C-NNNN"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.implementation ?? ""}
-                  onChange={(e) => onUpdate(key, { implementation: e.target.value || undefined })}
-                  onBlur={onCommit}
-                  placeholder="PG sequence + DEFAULT"
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: customerCode)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !numbering[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(numbering, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-function TxEditor({
-  tx, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  tx: Record<string, TxEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<TxEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(tx);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "14em" }} />
-          <col />
-          <col style={{ width: "18em" }} />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.tx.xxx)</th>
-            <th>policy</th>
-            <th>description</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <textarea
-                  className="form-control form-control-sm conventions-table-textarea"
-                  value={entry.policy}
-                  onChange={(e) => onUpdate(key, { policy: e.target.value })}
-                  onBlur={onCommit}
-                  rows={2}
-                  placeholder="単一操作は 1 TX..."
-                />
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: singleOperation)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !tx[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(tx, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
-  );
-}
-
-const OUTCOME_OPTIONS: ExternalOutcomeEntry["outcome"][] = ["success", "failure", "timeout"];
-const ACTION_OPTIONS: ExternalOutcomeEntry["action"][] = ["continue", "abort", "compensate"];
-const RETRY_OPTIONS = ["", "none", "fixed", "exponential"] as const;
-
-function ExternalOutcomeDefaultEditor({
-  entries: entriesMap, onAdd, onUpdate, onCommit, onRemove, isReadonly,
-}: {
-  entries: Record<string, ExternalOutcomeEntry>;
-  onAdd: (key: string) => void;
-  onUpdate: (key: string, patch: Partial<ExternalOutcomeEntry>) => void;
-  onCommit: () => void;
-  onRemove: (key: string) => void;
-  isReadonly?: boolean;
-}) {
-  const [newKey, setNewKey] = useState("");
-  const entries = Object.entries(entriesMap);
-
-  return (
-    <EntriesWrapper empty={entries.length === 0}>
-      <table className="conventions-table">
-        <colgroup>
-          <col style={{ width: "12em" }} />
-          <col style={{ width: "10em" }} />
-          <col style={{ width: "10em" }} />
-          <col style={{ width: "12em" }} />
-          <col />
-          <col style={{ width: 28 }} />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>key (@conv.externalOutcomeDefaults.xxx)</th>
-            <th>outcome</th>
-            <th>action</th>
-            <th>retry</th>
-            <th>description</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {entries.map(([key, entry]) => (
-            <tr key={key}>
-              <td><code className="conventions-key-badge">{key}</code></td>
-              <td>
-                <select
-                  className="form-select form-select-sm"
-                  value={entry.outcome}
-                  onChange={(e) => { onUpdate(key, { outcome: e.target.value as ExternalOutcomeEntry["outcome"] }); onCommit(); }}
-                >
-                  {OUTCOME_OPTIONS.map((o) => <option key={o} value={o}>{o}</option>)}
-                </select>
-              </td>
-              <td>
-                <select
-                  className="form-select form-select-sm"
-                  value={entry.action}
-                  onChange={(e) => { onUpdate(key, { action: e.target.value as ExternalOutcomeEntry["action"] }); onCommit(); }}
-                >
-                  {ACTION_OPTIONS.map((o) => <option key={o} value={o}>{o}</option>)}
-                </select>
-              </td>
-              <td>
-                <select
-                  className="form-select form-select-sm"
-                  value={entry.retry ?? ""}
-                  onChange={(e) => {
-                    onUpdate(key, { retry: (e.target.value || undefined) as ExternalOutcomeEntry["retry"] });
-                    onCommit();
-                  }}
-                >
-                  {RETRY_OPTIONS.map((o) => <option key={o} value={o}>{o || "(未指定)"}</option>)}
-                </select>
-              </td>
-              <td>
-                <input
-                  className="form-control form-control-sm"
-                  value={entry.description ?? ""}
-                  onChange={(e) => onUpdate(key, { description: e.target.value || undefined })}
-                  onBlur={onCommit}
-                />
-              </td>
-              <td className="text-center"><DeleteBtn onClick={() => onRemove(key)} isReadonly={isReadonly} /></td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <NewKeyRow
-        placeholder="新規 key (例: failure)"
-        value={newKey}
-        setValue={setNewKey}
-        onAdd={() => { const k = newKey.trim(); if (k && !entriesMap[k]) { onAdd(k); setNewKey(""); } }}
-        disabled={!newKey.trim() || Object.hasOwn(entriesMap, newKey.trim())}
-        isReadonly={isReadonly}
-      />
-    </EntriesWrapper>
   );
 }
